FEAT: Get City from GPS
Takes the GPS info from meta_dict and converts it from DMS to DD format.

(35.0, 41.0, 37.9) to 35.693861

I then use the Nominatim API to get the address. Using their zoom level of 10 I can make the result start with the city, then split off the text before ", " to only get the city.

During development I learned that Nominatim will return the address in the local language if the language is not specified.