FIX: Added logic to handle missing metadata
Added some logic to handle when the file does not have metadata.

If there is no metadata, or it is missing the date I want the file to remain untouched. If it is missing GPS data I want the file to be renamed and moved anyway.

I also set up a basic log so that I can see why files are missing info or not being renamed. I want to expand this to log any other errors.