FIX: Added error handling when getting the metadata
I don't want the code to error and stop if a file cannot be processed. Added try statement to get_metadata so if pillow cannot open the file as an image, or if the image type does not have exif data it just returns an empty meta_dict and the error handling already present in main will skip the file and log it.

I was also able to bring get_metadata to only 3 indentations! This makes it Linux code style compliant ;)

"if you need more than 3 levels of indentation, you're screwed anyway, and should fix your program."

https://www.kernel.org/doc/html/latest/process/coding-style.html